maint: Major refactoring
This is a commit of a major refactoring in the project. Céu-Media library
has been restructures for promoting its extension. In this commit I've adapted
the project to accommodate Céu-Media restructuring.

Besides, the include folder was also restructured. The directory include/mars
contains the core sources of the project, while the directory include/sync
contains the synchronizers implemented (currently there is only mirror.ceu).

* Makefile
* include/mars-media.ceu
* include/mars-client.ceu -> include/mars/mars-client.ceu
* include/mars-defs.ceu -> include/mars/mars-defs.ceu
* include/mars/mars-media.ceu
* include/mars-server.ceu -> include/mars/mars-server.ceu
* include/mirror.ceu -> include/sync/mirror.ceu
* src/player.ceu
* src/server.ceu